+++ rslang/src/pages/Audio/components/AudioGreetings.interface.ts
@@ -1,0 +1,6 @@
+import { MouseEvent } from 'react';
+
+export interface IGreetings {
+  onClick: (event: MouseEvent<HTMLButtonElement>) => void;
+  isGameOpenFromMenu: boolean;
+}
