+++ rslang/src/pages/Audio/components/AudioStatistics/AudioStatistics.interface.ts
@@ -1,0 +1,12 @@
+export interface IAudioStatistics {
+  statistics: {
+    id: string;
+    word: string;
+    audio: string;
+    transcription: string;
+    wordTranslate: string;
+    result: boolean;
+  }[];
+  endGame: () => void;
+  resetGame: () => void;
+}
