+++ rslang/src/pages/Audio/components/buttons/buttonLevel/ButtonLevel.interface.ts
@@ -1,0 +1,7 @@
+import { MouseEvent } from 'react';
+
+export interface IButtonLevel {
+  description: string;
+  bgColor: string;
+  onClick: (event: MouseEvent<HTMLButtonElement>) => void;
+}
