+++ rslang/src/pages/Audio/components/buttons/buttonReset/ButtonReset.interface.ts
@@ -1,0 +1,6 @@
+export interface IButtonReset {
+  description: string;
+  bgColor?: string;
+  disabled?: boolean;
+  onClick: () => void;
+}
