+++ rslang/src/pages/Audio/components/buttons/buttonSelect/ButtonSelect.interface.ts
@@ -1,0 +1,8 @@
+import { MouseEvent } from 'react';
+
+export interface IButtonSelect {
+  description: string;
+  bgColor: string;
+  onClick: (event: MouseEvent<HTMLButtonElement>) => void;
+  disabled: boolean;
+}
