+++ rslang/src/pages/Audio/components/buttons/buttonSpeak/ButtonSpeak.interface.ts
@@ -1,0 +1,3 @@
+export interface IButtonSpeak {
+  audioLink?: undefined | string;
+}
