+++ rslang/src/pages/Audio/components/buttons/buttonSpeak/ButtonSpeak.tsx
@@ -1,11 +1,8 @@
 import { FC } from 'react';
 import cat from '../../../../../images/cat-speak.svg';
 import { WordsType } from '../../../constants';
+import { IButtonSpeak } from './ButtonSpeak.interface';
 import './ButtonSpeak.scss';
-
-export interface IButtonSpeak {
-  audioLink?: undefined | string;
-}
 
 const ButtonSpeak: FC<IButtonSpeak> = ({ audioLink }) => {
   const handleAudio = () => {
