+++ rslang/src/pages/Audio/components/circle/Cicle.interface.ts
@@ -1,0 +1,4 @@
+export interface ICircle {
+  title: string;
+  value: string | number;
+}
