+++ rslang/src/pages/Audio/components/circle/Circle.tsx
@@ -1,9 +1,5 @@
 import { FC } from 'react';
-
-export interface ICircle {
-  title: string;
-  value: string | number;
-}
+import { ICircle } from './Cicle.interface';
 
 const Circle: FC<ICircle> = ({ title, value }) => {
   return (
