+++ rslang/src/pages/Audio/components/game/Game.interface.ts
@@ -1,0 +1,16 @@
+import { WordsType, IStatistics } from '../../constants';
+
+export interface IGame {
+  data: WordsType[] | undefined;
+  group: number;
+  handleStatistics: ({
+    id,
+    audio,
+    word,
+    wordTranslate,
+    transcription,
+    result,
+  }: IStatistics) => void;
+  resetGame: () => void;
+  handleIsEndGame: (value: boolean) => void;
+}
