+++ rslang/src/pages/Audio/components/multiplier/Multiplier.interface.ts
@@ -1,0 +1,5 @@
+export interface IMultiplier {
+  multiplier: number;
+  value: number;
+  description: string;
+}
