+++ rslang/src/pages/Audio/components/tableRow/TableRow.interface.ts
@@ -1,0 +1,7 @@
+export interface ITableRow {
+  word: string;
+  audio: string;
+  transcription: string;
+  wordTranslate: string;
+  result: boolean;
+}
